+++ sqlite/changelog.md
@@ -6,10 +6,10 @@
   * SQLiteを閉じる機能
   * 変数の型からSQLiteのTableを作成する機能
   * SQLiteのTableを削除する機能
+  * SQLiteのTableからデータを読み取る機能
 
 # To Do
 * SQLiteのTableで追加があるならば変更する機能
-* SQLiteのTableからデータを読み取る機能
 * SQLiteのTableから検索する機能(オプションあり)
 * SQLiteのTableからデータを追加する機能
 * SQLiteのTableからid指定によるデータを更新機能